Cannot lose on first turn
Added logic to prevent the player from instantly losing.
New boolean "hasPlayerMadeFirstMove" added to GridManager. Each tile checks if this is false, if so, it checks if it is flagged or not. If it isn't, then when the player hits the tile, if it was a mine, it instead reassigns that mine elsewhere and recalculates the values (for both the clicked tile and the "new" mine tile.)