Battleship Bones
Added the bones for implementing battleship.
Added a tile created in blender to represent the tiles in the game
Added logic to control gamestate via the abstract class BattleshipGameState.
Added classes for battleship tiles and Ships, as well as enums to handle what ships the players are to be given.