Battleship Multiplayer
Works locally, build only the NetworkBattleship scene and run it separately as an exe. In the editor, run the game and hit "Host", and in the exe hit "Client" to begin the game. It doesn't seem like you can finish a game due to a bug but it mostly works