AI Skeleton pretty much working
We implemented the decision matrix. We separated the hit matrix into its own thing, and smarter AI prefers to focus on the hit matrix. The AI is what I would call "marginally competent" at the moment.

One more obvious improvement i could still make is to prefer shooting in the direction of chains of hits